Rethink the lt dependency_libs logic with the clarity of having slept on it
This avoids passing extra .la arguments to CCLD; lt.lib.post is the ONLY
variable that needs to know the recursive dependencies (so change
lt.{lib,exe}.files.all to not expand them); for the others,
libtool will figure it out from dependency_libs.  Further, if the library
is public (lt.lib.rpath is set), then we DON'T need to adjust
depedency_libs, as we've already told libtool about them in
lt.lib.files.ld.

This also changes how lt.lib.files.ld is implemented.  This is just for
readability, the old and new definitions are equivalent (except for
possible whitespace differences?).