Run the new fixup script.
The biggest change is that it keeps looking for new #include
blocks.  This means that it (more) correctly handles #ifdef'd
#includes.

I'm not 100% in love with it, but it's pretty good for automated.
What I really don't like is it did some silly things with newlines
in typedef blocks.  I've avoided committing those.  I think it may
be possible to get it to do the right thing.  But really, the
typedef blocks are a hack for poorly structured headers.

Now that it keeps looking for new #include blocks, I could strip
out the typedef support, and it would mostly work; but it wouldn't
be able to move some headers back to the top.

IDK what to do.