tools/notsd-fixup: Split the makefile pass into a separate executable.
This avoids using bash read(1) to loop over the files, as we've already gotten
decent speedups from avoiding bash read(1).  This means we've got at least two
more fork/exec's because of xargs, but it's probably worth it.

But that's kind of premature-optimization; the time improvement here is
probably just random noise.

But, I think this makes the code more maintainable/manageable too, so I'm
committing it.