nspawn: mount_sysfs(): Reword the comment about /sys/fs/cgroup
The comment explains the obvious, but doesn't even mention the tricky part.
Of course we need do set things up before we remount read-only!  That's
the general theme of the function!

What was totally non-obvious is why we only need to create it if
cg_ns_supported(), as the directory needs to exist no matter what.  From
reading the code, I was convinced that it was broken on pre-cgns  kernels
(pre-4.6, unless a distro backported it).

So explain that skippint creating if !cg_ns_supported() is an optimization.