nspawn: mount_legacy_cgns_supported(): Rename variables to not lie
mount_legacy_cgns_supported() is very clearly meant to be a version of
mount_legacy_cgns_unsupported() modified to cope with the fact that it has
already chroot()ed, and thus can't look at the host /sys.  So, the loops
and such look similar.

However, to cope with the fact that it can't look at /sys, it deals with
hierarchies in the outermost loop, rather than controllers.  Yet, it kept
the list variable named "controllers".  That's confusing.