cgroup-util: Split out cg_pid_get_path_internal()
The name is a bit if a misnomer, because it doesn't work with a PID, but
with an already-opened /proc/${pid}/cgroup 'FILE*'.

It lets us bypass the controller=NULL and
controller=SYSTEMD_CGROUP_CONTROLLER special cases of cg_pid_get_path(), so
that we can directly check the cgroup v2 hierarchy; even if systemd isn't
using it (useful if we're doing work with a container that does use the v2
hierarchy).  It also requires the caller to have already fopen()ed the
cgroup file (useful so that we can read from an fd passed over a socket).