nspawn: Add functions for deciding cgroup mounts before performing them
This is part 1 of a 2-part commit; it adds the functions, but doesn't
modify anything to use them; that's part 2.  I've split it up for clarity
of the diff, and to make rebasing easier.

Add (static) cgmount_add(), to build a list of CGMounts;
cgroup_free_mounts() to eventually free that list; and
cgroup_mount_mounts() to perform the mounts in the list.

The behavior of cgroup_mount_mounts() borrows from/imitates:

 - mount_legacy_cgroup_hierarchy(): Most everything, but it shoves the
   decision of whether to use cgroup v1 or v2 to be the responsibility of
   whoever is building the list.
 - mount_legacy_cgns_[un]supported(): The tmpfs logic, the logics on when
   to mount a cgroup hierarchy RO or RW, the logics on when to remount
   /sys/fs/cgroup RO.