nspawn: Go ahead and always decide the cgroup mounts in the outer child, not inner
It mounts them in the outer child if !use_cgns, or inner child if use_cgns.
Previously, it decided the mounts at the same place that it actually
mounted them.  But, for simplicity and flexibility, always decide them in
the outer child.

Also, drop the now superfluous "dest" argument to cgroup_mount_mounts().