nspawn: Change where we filter the name=systemd hierarchy
It was being filtered out in get_v1_hierarchies(), which seems a little
silly because it is a v1 hierarchy.  Instead, filter it in
mount_legacy_cgns_supported(); the function that actually cares about
skipping it.  This is a little less efficient, since get_v1_hierarchies()
now has to do the allocations to put it in the returned set; but that's
minor and the code is clearer now.