cgroup: Merge the unified_cache and unified_systemd_v232 caches
Conceptually, the addition of bool unified_systemd_v232 split
CGROUP_UNIFIED_SYSTEMD in to two separate values.  So, split it.

While we're at it, rename the CGroupUnified values.  Before we could think
of it as "how much of the system is unified?", but with us now caring about
exactly which mountpoint is unified, I don't think it's a helpful way of
conceptualizing it when reasoning about the code.

 - CGROUP_UNIFIED_UNKNOWN --> CGROUP_VER_UNKNOWN

 - CGROUP_UNIFIED_NONE    --> CGROUP_VER_1_SD

     Saying "no cgroup hierarchies are unified" is the same as saying "all
     cgroup hierarchies are cgroup v1", so let's name it "CGROUP_VER_1".
     But that's not quite descriptive enough, as there are infinitely many
     cgroup v1 configurations (one of the problems with cgroup v1); and to
     us CGROUP_UNIFIED_NONE meant something a bit more specific: "cgroup v1
     as implemented by systemd that is, with a "name=systemd" hierarchy
     mounted at /sys/fs/cgroup/systemd.  So add an "_SD" suffix to convey
     the systemd-specificity of this arrangemnt.

 - CGROUP_UNIFIED_ALL     --> CGROUP_VER_2

     Saying "all cgroup hierarchies are unified" is the same as saying "we
     are using cgroup v2".

That's simple, enough; but what about the in-betweens, when we want to mix
v1 and v2?

 - CGROUP_UNIFIED_SYSTEMD --> CGROUP_VER_MIXED_SD232
                          `-> CGROUP_VER_MIXED_SD233

     What CGROUP_UNIFIED_SYSTEMD said is "we're mostly using v1 cgroup
     hierarchies, but are using the v2 hierarchy for systemd."  The
     specifics beyond that were conveyed by the separate
     `unified_systemd_v232` bool.  So, the suffix after "MIXED" conveys the
     coexistence arrangement between the v1 and v2 hierarchies.

     _SD232 conveys the arrangement used by systemd 232; that is, the v2
     (unified) hierarchy is mounted at /sys/fs/cgroup/systemd.

     _SD233 conveys the arrangement used by systemd 233 (and presumably
     future systemd versions); that is, a "name=systemd" v1 hierarchy is
     mounted at /sys/fs/cgroup/systemd, and the v2 (unified) hierarchy is
     mounted at /sys/fs/cgroup/unified; and systemd tries to keep them in
     sync.

There are some places that use comparisons like ">= CGROUP_UNIFIED_SYSTEMD"
when from the calling context we know that it is not CGROUP_UNIFIED_ALL; so it
might as well be "==" instead of ">=".  If it makes the code clearer, change
that to an "==" comparison.