nspawn: get_v1_hierarchies(): Ditch a pointless check for "name=unified"
name=unified isn't a real v1 hierarchy, it's just a made-up name to refer
to the v2 hierarchy with a v1-looking string.  It will never show up when
we ask the kernel for a list of v1 hierarchies.  It looks like Tejun Heo
just grepped for all uses of "name=systemd" when he introduced name=unified
in 2977724b09eb997fc84a80517447b5d4a70770c7.