nspawn: Decide all cgroup mounts/symlinks before performing any of them
This is part 2 of a 2-part commit; it modifies the pre-existing code to use
the cgmount_add(), cgroup_mount_mounts(), and cgroup_free_mounts()
functions added in part 1.

Instead of actually doing anything when making decisions around cgroup, we
use cgroup_decide_mounts_*() to build up a CGMounts structure that is a
list of actions to perform when setting up group in the container.  Then we
pass that to cgroup_mount_mounts(), which actually does everything.