cgroup-util,nspawn: Use switch cases around CGroupUnified when possible
This replaces a bunch of confusing if/else trees with
simple-to-reason-about switch blocks that make it absolutely clear what
happens in each case of the enum.

In nspawn, avoid all of the cg_* functions that query properties of the
underlying CGroupUnified, and just check cg_version directly.  Even in the
cases where the checks on it can't be done in a switch statement, it's
clearer, because it shows the symmetry between inner_cgver and outer_cgver.

This should not change the behavior at all; except that there are a few
more assert()s on things that shouldn't happen.