cgroup-util,nspawn: Add a special "inherit" cgroup mode for nspawn
The "inherit" mode inspects /proc/self/mountinfo to do its best to
replicate the cgroup setup of the outer host.  It is used by default unless
a different specific cgroup setup is to be used; either because the user
requested it (via $UNIFIED_CGROUP_HIERARCHY), or because
pick_cgroup_version() sniffed that the container has a version of systemd
that doesn't support the outer host's setup.

This means that nspawn can now be used when
outer_cgver=CGROUP_UNIFIED_UNKNOWN; AKA when running on a non-systemd host.